requirements.txt
streamlit
boto3
pyngrok
python-dotenv
python-docx
debugpy